Move packages to top-level aedifix directory
